Configure non-admin k8s user

+++ populate_pass_secrets.bash
@@ -19,18 +19,50 @@ function pass_generate_sealed_secrets_cert_if_absent
     local key_path="${temp_dir}/key"
     local cert_path="${temp_dir}/cert"
 
-    openssl req \
+    quiet openssl req \
       -x509 \
       -nodes \
       -newkey "rsa:4096" \
       -keyout "${key_path}" \
       -out "${cert_path}" \
-      -subj "/CN=sealed-secret/O=sealed-secret" &>/dev/null
+      -subj "/CN=sealed-secret/O=sealed-secret"
 
     cat "${key_path}" |
       pass_store "${secret_key_id}"
     cat "${cert_path}" |
       pass_store "${certificate_id}"
+
+    rm -rf "${temp_dir}"
+  fi
+}
+
+function pass_generate_k8s_user_csr_if_absent
+{
+  local username="$1"
+  local key_id="$2"
+  local csr_id="$3"
+
+  if ! pass_exists "${key_id}" ||
+    ! pass_exists "${csr_id}"; then
+    local temp_dir
+    temp_dir="$(mktemp -d)"
+    local key_path="${temp_dir}/key"
+    local csr_path="${temp_dir}/csr"
+
+    quiet openssl genrsa \
+      -out "${key_path}" \
+      4096
+    quiet openssl req \
+      -new \
+      -key "${key_path}" \
+      -out "${csr_path}" \
+      -sha256 \
+      -subj "/CN=${username}/O=${username}"
+
+    cat "${key_path}" |
+      pass_store "${key_id}"
+    cat "${csr_path}" |
+      pass_store "${csr_id}"
 
     rm -rf "${temp_dir}"
   fi
@@ -78,6 +110,11 @@ function main
   pass_generate_sealed_secrets_cert_if_absent \
     "${PASS_SECRET_ID_SEALED_SECRETS_KEY_kxlsnqam}" \
     "${PASS_SECRET_ID_SEALED_SECRETS_CERTIFICATE_4edcp3cm}"
+
+  pass_generate_k8s_user_csr_if_absent \
+    "wkaluza" \
+    "${PASS_SECRET_ID_K8S_USER_KEY_nzgamwny}" \
+    "${PASS_SECRET_ID_K8S_USER_CSR_xf7rrqr3}"
 
   gpg \
     --export-ssh-key "${primary_key_fingerprint}" |
